added noise generation

+++ internal/domain/audio_phrase.go
@@ -7,5 +7,5 @@ type AudioPhrase struct {
 	PathToAudio string    `json:"path_to_audio"`
 	PhraseID    uuid.UUID `json:"phrase_id"`
 	Accent      string    `json:"accent"`
-	Noise       int       `json:"noise"`
+	Noise       float64   `json:"noise"`
 }
